refactor: 마이페이지 전체 구조 분리 (#40)
